v1.1更新日志 1. 优化代码 2. 新增放置检测，若放置失败，则会自动向周边点位略微调整
v1.1 Update Log
Optimized code
Added placement detection; if placement fails, it will automatically adjust slightly to surrounding points.
