+++ BTD6AutoCommunity/BTD6AutoCommunity.csproj
@@ -108,6 +108,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Contants.cs" />
     <Compile Include="DisPlayMouseCoordinates.cs" />
     <Compile Include="ExecuteDirectiveClass.cs" />
     <Compile Include="Form1.cs">
@@ -122,14 +123,28 @@
     <Compile Include="Form2.Designer.cs">
       <DependentUpon>Form2.cs</DependentUpon>
     </Compile>
+    <Compile Include="GameStateMachine.cs" />
     <Compile Include="GetGameData.cs" />
     <Compile Include="HeroClass.cs" />
+    <Compile Include="HotKeys.cs" />
     <Compile Include="InstructionsClass.cs" />
     <Compile Include="KeyEvents.cs" />
     <Compile Include="MonkeyTowerClass.cs" />
+    <Compile Include="MyScriptsPage.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ScriptsEditer.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="SettingPage.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Settings.cs" />
+    <Compile Include="StartPage.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
